Crate multicore_bringup

Initialization and bring-up of secondary CPUs.

These functions are intended to be invoked from the BSP (the Bootstrap Processor, the main CPU in x86 terms) in order to bring up secondary CPUs (APs in x86 terms).

Structs

GraphicInfo

A structure to access information about the graphical framebuffer mode that was discovered and chosen in the AP’s real-mode initialization sequence.

MulticoreBringupInfo

Information needed to bring up APs (secondary CPUs) on x86_64.

Functions

get_graphic_info

Returns information about the currently-active graphical framebuffer.

handle_ap_cores

Starts up and sets up AP cores based on system information from ACPI (specifically the MADT (APIC) table).