Crate multicore_bringup

Functions for initializing and bringing up other CPU cores.

These functions are intended to be invoked from the main core (the BSP – bootstrap core) in order to jumpstart other cores.

Structs

GraphicInfo

A structure to access information about the graphical framebuffer mode that was discovered and chosen in the AP’s real-mode initialization sequence.

Statics

GRAPHIC_INFO

Graphic mode information that will be updated after handle_ap_cores() is invoked.

Functions

handle_ap_cores

Starts up and sets up AP cores based on system information from ACPI (specifically the MADT (APIC) table).